requirements.txt

asgiref==3.4.1
Django==3.2.11
django-appconf==1.0.5
django-bootstrap3==21.1
django-core==1.4.1
django-select2==7.9.0
djangorestframework==3.13.1
idna==3.3
importlib-metadata==2.1.2
install==1.3.5
pkg_resources==0.0.0
psycopg2==2.7.7
psycopg2-binary==2.9.3
pytz==2021.3
sqlparse==0.4.2
typing_extensions==4.0.1
urllib3==1.26.8
zipp==3.6.0
